CORE-1338 Add Launch Preview to App Editor
