+++ src/components/apps/editor/index.js
@@ -9,16 +9,17 @@ import { FastField, Formik } from "formik";
 
 import { useTranslation } from "i18n";
 
+import { formatSubmission, initAppValues } from "./formatters";
 import ids from "./ids";
 import styles from "./styles";
 
 import CmdLineOrderForm from "./CmdLineOrderForm";
 import ParamGroups from "./ParamGroups";
+import ParametersPreview from "./ParametersPreview";
 
 import AppStepper, { StepperSkeleton } from "../AppStepper";
 import AppStepDisplay, { BottomNavigationSkeleton } from "../AppStepDisplay";
 
-import AppParamTypes from "components/models/AppParamTypes";
 import SaveButton from "components/utils/SaveButton";
 import useComponentHeight from "components/utils/useComponentHeight";
 
@@ -43,246 +44,6 @@ import {
 import { ArrowBack, ArrowForward } from "@material-ui/icons";
 
 const useStyles = makeStyles(styles);
-
-/**
- * Initializes the form values from the given App Description.
- *
- * Removes the optional `name` field from groups,
- * initializes each text parameter's `defaultValue` to an empty string,
- * and each flag's `name` field to a custom object for this form.
- *
- * @param {Object} appDescription
- * @param {Object[]} appDescription.groups
- * @param {Object[]} appDescription.groups.parameters
- *
- * @returns Initial form values.
- */
-const initValues = (appDescription) => {
-    const { groups } = appDescription;
-
-    const initializedGroups = groups?.map(({ name, ...group }) => ({
-        ...group,
-        parameters: group.parameters?.map((param) => {
-            const {
-                name,
-                defaultValue,
-                type: paramType,
-                arguments: paramArgs,
-            } = param;
-
-            switch (paramType) {
-                case AppParamTypes.INTEGER:
-                    return {
-                        ...param,
-                        defaultValue:
-                            defaultValue || defaultValue === 0
-                                ? Number.parseInt(defaultValue)
-                                : "",
-                    };
-
-                case AppParamTypes.DOUBLE:
-                    return {
-                        ...param,
-                        defaultValue:
-                            defaultValue || defaultValue === 0
-                                ? Number.parseFloat(defaultValue)
-                                : "",
-                    };
-
-                case AppParamTypes.FLAG:
-                    return {
-                        ...param,
-                        name: initFlagName(name),
-                        defaultValue: defaultValue && defaultValue !== "false",
-                    };
-
-                case AppParamTypes.FILE_INPUT:
-                case AppParamTypes.FOLDER_INPUT:
-                    return {
-                        ...param,
-                        defaultValue: defaultValue?.path || "",
-                    };
-
-                case AppParamTypes.MULTIFILE_SELECTOR:
-                    return {
-                        ...param,
-                        defaultValue: defaultValue?.path || [],
-                    };
-
-                default:
-                    let defaultArg;
-                    if (paramArgs?.length > 0) {
-                        defaultArg =
-                            paramArgs.find(
-                                (arg) => defaultValue?.id === arg.id
-                            ) || paramArgs.find((arg) => arg.isDefault);
-                    }
-
-                    return {
-                        ...param,
-                        defaultValue: defaultArg || defaultValue || "",
-                    };
-            }
-        }),
-    }));
-
-    return { ...appDescription, groups: initializedGroups };
-};
-
-/**
- * @typedef {object} FlagNameModelOptVal
- * @property {string} option
- * @property {string} value
- */
-
-/**
- * @typedef {object} FlagNameModel
- * @property {FlagNameModelOptVal} checked
- * @property {FlagNameModelOptVal} unchecked
- */
-
-/**
- * Parses `Flag` param `name` strings into a custom object for this form.
- *
- * @param {string} name
- *
- * @returns {FlagNameModel} An object for use in the flag property editor form.
- */
-const initFlagName = (name) => {
-    const [checked, unchecked] = name?.split(", ") || [];
-
-    const [checkedOpt, ...checkedVal] = checked?.split(" ") || [];
-    const [uncheckedOpt, ...uncheckedVal] = unchecked?.split(" ") || [];
-
-    return {
-        checked: {
-            option: checkedOpt || "",
-            value: checkedVal.join(" "),
-        },
-        unchecked: {
-            option: uncheckedOpt || "",
-            value: uncheckedVal.join(" "),
-        },
-    };
-};
-
-/**
- * Formats the form values for submission to the service.
- *
- * @param {Object} appDescription
- * @param {Object[]} appDescription.groups
- * @param {Object[]} appDescription.groups.parameters
- *
- * @returns The App Description formatted for submission to the service.
- */
-const formatSubmission = (appDescription) => {
-    const { groups } = appDescription;
-
-    const formattedGroups = groups?.map((group) => ({
-        ...group,
-        parameters: group.parameters?.map((param) => {
-            const {
-                name,
-                defaultValue,
-                type: paramType,
-                arguments: paramArgs,
-            } = param;
-
-            switch (paramType) {
-                case AppParamTypes.TEXT:
-                case AppParamTypes.MULTILINE_TEXT:
-                case AppParamTypes.REFERENCE_ANNOTATION:
-                case AppParamTypes.REFERENCE_GENOME:
-                case AppParamTypes.REFERENCE_SEQUENCE:
-                    return {
-                        ...param,
-                        defaultValue: defaultValue || null,
-                    };
-
-                case AppParamTypes.INTEGER:
-                case AppParamTypes.DOUBLE:
-                    return {
-                        ...param,
-                        defaultValue:
-                            defaultValue || defaultValue === 0
-                                ? defaultValue
-                                : null,
-                    };
-
-                case AppParamTypes.FLAG:
-                    return {
-                        ...param,
-                        name: formatFlagName(name),
-                        defaultValue: defaultValue && defaultValue !== "false",
-                    };
-
-                case AppParamTypes.TEXT_SELECTION:
-                case AppParamTypes.INTEGER_SELECTION:
-                case AppParamTypes.DOUBLE_SELECTION:
-                    return {
-                        ...param,
-                        arguments: formatSelectionArgs(paramArgs, defaultValue),
-                        defaultValue: defaultValue
-                            ? { ...defaultValue, isDefault: true }
-                            : null,
-                    };
-
-                case AppParamTypes.FILE_INPUT:
-                case AppParamTypes.FOLDER_INPUT:
-                    return {
-                        ...param,
-                        defaultValue: defaultValue
-                            ? { path: defaultValue }
-                            : null,
-                    };
-
-                case AppParamTypes.MULTIFILE_SELECTOR:
-                    // default values not yet supported
-                    return {
-                        ...param,
-                        defaultValue: null,
-                    };
-
-                default:
-                    return param;
-            }
-        }),
-    }));
-
-    return { ...appDescription, groups: formattedGroups };
-};
-
-/**
- * Formats this form's custom `Flag` param `name` model
- * into a string expected by the service.
- *
- * @param {FlagNameModel} name
- *
- * @returns A string in the form
- *          "checked.option checked.value, unchecked.option unchecked.value"
- */
-const formatFlagName = (name) => {
-    const {
-        checked: { option: checkedOpt, value: checkedVal },
-        unchecked: { option: uncheckedOpt, value: uncheckedVal },
-    } = name;
-
-    return [
-        [checkedOpt, checkedVal].join(" ").trim(),
-        [uncheckedOpt, uncheckedVal].join(" ").trim(),
-    ].join(", ");
-};
-
-const formatSelectionArgs = (paramArgs, defaultValue) => {
-    return paramArgs?.map((paramArg) => ({
-        ...paramArg,
-        isDefault:
-            paramArg.id === defaultValue?.id ||
-            (paramArg.display === defaultValue?.display &&
-                paramArg.name === defaultValue?.name &&
-                paramArg.value === defaultValue?.value),
-    }));
-};
 
 const displayStepError = (stepIndex, errors, touched) => {
     if (stepIndex === 0) {
@@ -346,16 +107,16 @@ const AppEditor = (props) => {
         label: t("parameters"),
         contentLabel: t("appParameters"),
     };
+    const stepPreview = {
+        label: t("previewApp"),
+        contentLabel: t("previewApp"),
+    };
     const stepCmdLineOrder = {
         label: t("commandLineOrder"),
         contentLabel: t("commandLineOrder"),
     };
-    const stepPreview = {
-        label: t("previewApp"),
-        contentLabel: t("common:comingSoon"),
-    };
 
-    const steps = [stepAppInfo, stepParameters, stepCmdLineOrder, stepPreview];
+    const steps = [stepAppInfo, stepParameters, stepPreview, stepCmdLineOrder];
 
     const isLastStep = () => {
         return activeStep === steps.length - 1;
@@ -382,7 +143,7 @@ const AppEditor = (props) => {
     return (
         <Formik
             enableReinitialize
-            initialValues={initValues({ ...appDescription })}
+            initialValues={initAppValues({ ...appDescription })}
             validate={(values) => {
                 const errors = {};
 
@@ -513,6 +274,11 @@ const AppEditor = (props) => {
                                     baseId={baseId}
                                     groups={values.groups}
                                 />
+                            ) : activeStepInfo === stepPreview ? (
+                                <ParametersPreview
+                                    baseId={baseId}
+                                    groups={values.groups}
+                                />
                             ) : activeStepInfo === stepCmdLineOrder ? (
                                 <CmdLineOrderForm
                                     baseId={baseId}
@@ -520,8 +286,6 @@ const AppEditor = (props) => {
                                     groups={values.groups}
                                     setFieldValue={setFieldValue}
                                 />
-                            ) : activeStepInfo === stepPreview ? (
-                                t("common:comingSoon")
                             ) : null}
                         </AppStepDisplay>
                     </Paper>
